Introduce helper.get_scaled_size(size: int) -> int
This function uses settings["ui_scale"][1], but it would have to load the settings file every time we use it. We need to store the loaded settings somewhere.

We are using an attribute of the function, that is, we are checking if get_scaled_size.settings exists, and if it does, we use this. If it does not exist, we load the settings with get_settings() and store it in this attribute.

This is a dirty solution (there should be a singleton class), but it works for now.